Conflict Free Updates
Updates to Update classes to deal with private functions.

Pull apart doc and state caching functionality and change private methods to protected.

Strip out state functionality from ConflictFreeUpdate
